Reorganize test utility functions and global test setup
Refactors the test utilities and global test setup by:
- Moving helper functions to a new `_utils` directory
- Creating a global test setup file for ModuleProxyFactory
- Updating import paths in test files to reflect the new structure
- Removing the old `helpers.ts` file

This change improves test code organization and makes utility functions more modular and accessible across test suites.
