+++ test/deployables/freeze-guard/AzoriusFreezeGuardV1.test.ts
@@ -7,8 +7,8 @@ import {
   MockFreezeVoting,
   MockFreezeVoting__factory,
 } from '../../../typechain-types';
-import { getModuleProxyFactory } from '../../global/GlobalSafeDeployments.test';
-import { calculateProxyAddress } from '../../helpers';
+import { getModuleProxyFactory } from '../../helpers/globals.test';
+import { calculateProxyAddress } from '../../helpers/utils';
 
 // Helper function for deploying AzoriusFreezeGuardV1 instances
 async function deployAzoriusFreezeGuardProxy(
