+++ test/deployables/modules/AzoriusV1.test.ts
@@ -12,8 +12,8 @@ import {
   MockVotingStrategy,
   MockVotingStrategy__factory,
 } from '../../../typechain-types';
-import { getModuleProxyFactory } from '../../global/GlobalSafeDeployments.test';
-import { calculateProxyAddress } from '../../helpers';
+import { getModuleProxyFactory } from '../../helpers/globals.test';
+import { calculateProxyAddress } from '../../helpers/utils';
 
 // Helper functions for deploying AzoriusV1 instances
 async function deployAzoriusProxy(
