+++ test/deployables/modules/FractalModuleV1.test.ts
@@ -9,8 +9,8 @@ import {
   MockERC20,
   MockERC20__factory,
 } from '../../../typechain-types';
-import { getModuleProxyFactory } from '../../global/GlobalSafeDeployments.test';
-import { calculateProxyAddress } from '../../helpers';
+import { getModuleProxyFactory } from '../../helpers/globals.test';
+import { calculateProxyAddress } from '../../helpers/utils';
 
 // Helper functions for deploying FractalModuleV1 instances
 async function deployFractalModuleProxy(
