+++ test/deployables/strategies/BaseQuorumPercentV1.test.ts
@@ -5,8 +5,8 @@ import {
   ConcreteBaseQuorumPercentV1,
   ConcreteBaseQuorumPercentV1__factory,
 } from '../../../typechain-types';
-import { getModuleProxyFactory } from '../../global/GlobalSafeDeployments.test';
-import { calculateProxyAddress } from '../../helpers';
+import { getModuleProxyFactory } from '../../helpers/globals.test';
+import { calculateProxyAddress } from '../../helpers/utils';
 
 describe('BaseQuorumPercentV1', () => {
   // Signers
