+++ test/deployables/strategies/BaseVotingBasisPercentV1.test.ts
@@ -5,8 +5,8 @@ import {
   ConcreteBaseVotingBasisPercentV1,
   ConcreteBaseVotingBasisPercentV1__factory,
 } from '../../../typechain-types';
-import { getModuleProxyFactory } from '../../global/GlobalSafeDeployments.test';
-import { calculateProxyAddress } from '../../helpers';
+import { getModuleProxyFactory } from '../../helpers/globals.test';
+import { calculateProxyAddress } from '../../helpers/utils';
 
 describe('BaseVotingBasisPercentV1', () => {
   // Signers
