+++ test/deployables/strategies/LinearERC721VotingV1.test.ts
@@ -9,8 +9,8 @@ import {
   MockERC721__factory,
   MockOwnership__factory,
 } from '../../../typechain-types';
-import { getModuleProxyFactory } from '../../global/GlobalSafeDeployments.test';
-import { calculateProxyAddress } from '../../helpers';
+import { getModuleProxyFactory } from '../../helpers/globals.test';
+import { calculateProxyAddress } from '../../helpers/utils';
 
 describe('LinearERC721VotingV1', () => {
   // Signers
